Split GUI settings handlers into controllers
